requirements.txt
benchmark-functions==1.1.3
contourpy==1.3.1
customtkinter==5.2.2
cycler==0.12.1
darkdetect==0.8.0
fonttools==4.56.0
kiwisolver==1.4.8
matplotlib==3.10.1
numpy==2.2.4
packaging==24.2
pillow==11.1.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
scipy==1.15.2
six==1.17.0
